+++ src/main/java/com/nisovin/shopkeepers/metrics/FeaturesChart.java
@@ -22,7 +22,7 @@ public class FeaturesChart extends Metrics.DrilldownPie {
 		super("used_features", () -> {
 			Map<String, Map<String, Integer>> allFeatures = new LinkedHashMap<>();
 			// plugin compatibility features:
-			addFeatureEntry(allFeatures, "bypass-shop-interaction-blocking", Settings.bypassShopInteractionBlocking);
+			addFeatureEntry(allFeatures, "check-shop-interaction-result", Settings.checkShopInteractionResult);
 			addFeatureEntry(allFeatures, "bypass-spawn-blocking", Settings.bypassSpawnBlocking);
 			addFeatureEntry(allFeatures, "enable-spawn-verifier", Settings.enableSpawnVerifier);
 			addFeatureEntry(allFeatures, "enable-world-guard-restrictions", Settings.enableWorldGuardRestrictions && WorldGuardHandler.isPluginEnabled());