+++ src/main/java/com/nisovin/shopkeepers/util/Utils.java
@@ -29,6 +29,10 @@ import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.EntityType;
 import org.bukkit.entity.Player;
+import org.bukkit.event.Event.Result;
+import org.bukkit.event.block.Action;
+import org.bukkit.inventory.ItemStack;
+import org.bukkit.inventory.PlayerInventory;
 import org.bukkit.permissions.Permissible;
 import org.bukkit.util.RayTraceResult;
 import org.bukkit.util.Vector;
@@ -36,6 +40,77 @@ import org.bukkit.util.Vector;
 public final class Utils {
 
 	private Utils() {
+	}
+
+	/**
+	 * Checks if the player can interact with the given block.
+	 * <p>
+	 * This works by clearing the player's items in main and off hand, calling a dummy PlayerInteractEvent for plugins
+	 * to react to and then restoring the player's items in main and off hand.
+	 * <p>
+	 * Since this involves calling a dummy PlayerInteractEvent, plugins reacting to the event might cause all kinds of
+	 * side effects. Therefore, this should only be used in very specific situations, such as for specific blocks.
+	 * 
+	 * @param player
+	 *            the player
+	 * @param block
+	 *            the block to check interaction with
+	 * @return <code>true</code> if no plugin denied block interaction
+	 */
+	public static boolean checkBlockInteract(Player player, Block block) {
+		// simulating right click on the block to check if access is denied:
+		// making sure that block access is really denied, and that the event is not cancelled because of denying
+		// usage with the items in hands:
+		PlayerInventory playerInventory = player.getInventory();
+		ItemStack itemInMainHand = playerInventory.getItemInMainHand();
+		ItemStack itemInOffHand = playerInventory.getItemInOffHand();
+		playerInventory.setItemInMainHand(null);
+		playerInventory.setItemInOffHand(null);
+
+		TestPlayerInteractEvent dummyInteractEvent = new TestPlayerInteractEvent(player, Action.RIGHT_CLICK_BLOCK, null, block, BlockFace.UP);
+		Bukkit.getPluginManager().callEvent(dummyInteractEvent);
+		boolean canAccessBlock = (dummyInteractEvent.useInteractedBlock() != Result.DENY);
+
+		// resetting items in main and off hand:
+		playerInventory.setItemInMainHand(itemInMainHand);
+		playerInventory.setItemInOffHand(itemInOffHand);
+		return canAccessBlock;
+	}
+
+	/**
+	 * Checks if the player can interact with the given entity.
+	 * <p>
+	 * This works by clearing the player's items in main and off hand, calling a dummy PlayerInteractEntityEvent for
+	 * plugins to react to and then restoring the player's items in main and off hand.
+	 * <p>
+	 * Since this involves calling a dummy PlayerInteractEntityEvent, plugins reacting to the event might cause all
+	 * kinds of side effects. Therefore, this should only be used in very specific situations, such as for specific
+	 * entities, and its usage should be optional (i.e. guarded by a config setting).
+	 * 
+	 * @param player
+	 *            the player
+	 * @param entity
+	 *            the entity to check interaction with
+	 * @return <code>true</code> if no plugin denied interaction
+	 */
+	public static boolean checkEntityInteract(Player player, Entity entity) {
+		// simulating right click on the entity to check if access is denied:
+		// making sure that entity access is really denied, and that the event is not cancelled because of denying usage
+		// with the items in hands:
+		PlayerInventory playerInventory = player.getInventory();
+		ItemStack itemInMainHand = playerInventory.getItemInMainHand();
+		ItemStack itemInOffHand = playerInventory.getItemInOffHand();
+		playerInventory.setItemInMainHand(null);
+		playerInventory.setItemInOffHand(null);
+
+		TestPlayerInteractEntityEvent dummyInteractEvent = new TestPlayerInteractEntityEvent(player, entity);
+		Bukkit.getPluginManager().callEvent(dummyInteractEvent);
+		boolean canAccessEntity = !dummyInteractEvent.isCancelled();
+
+		// resetting items in main and off hand:
+		playerInventory.setItemInMainHand(itemInMainHand);
+		playerInventory.setItemInOffHand(itemInOffHand);
+		return canAccessEntity;
 	}
 
 	public static String getServerCBVersion() {